feat(Browserversion): allow to import browser version
