+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "javascript-i18n-library",
-  "version": "1.6.0",
+  "version": "1.7.0",
   "description": "This javascript library is used to format dates, numbers and currencies. It's compatible with Node and Browsers.",
   "main": "javascript-i18n-library.js",
   "repository": {
